+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fe-fpm-cli
 
+## 1.0.18
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fe-fpm-writer 1.2.0 → 1.2.1
+
 ## 1.0.17
 
 ### Patch Changes