+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 1.0.22
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.0.20 → 1.1.0
+- @sap-ux/ui5-library-inquirer 1.0.21 → 1.0.22
+
 ## 1.0.21
 
 ### Patch Changes